+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ## 1.3.0
 
+- [issue #6] Better error handling for running with a Python that doesn't
+  have pyexpat installed (and only when needed for Jira SOAP API calls).
 - '-o' flag to 'jirash issues' to include open tickets, but NOT just "Open". Also
   include "Reopened" and "In Progress".
 - Fix case-insensitive searching of statuses in 'jirash issues -s ...'.